Clean up twistd.pid after a force kill

+++ master/buildbot/scripts/stop.py
@@ -42,7 +42,8 @@ def stop(config, signame="TERM", wait=False):
 
     signum = getattr(signal, "SIG"+signame)
     try:
-        os.kill(pid, signum)
+        if base.isBuildBotRunning(basedir, quiet):
+            os.kill(pid, signum)
     except OSError, e:
         if e.errno != errno.ESRCH:
             raise
